+++ model/src/test/java/denominator/model/profile/GeoTest.java
@@ -9,11 +9,7 @@ import org.testng.annotations.Test;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMultimap;
-
-import denominator.model.ResourceRecordSet;
-import denominator.model.rdata.AData;
 
 @Test
 public class GeoTest {
@@ -23,7 +19,7 @@ public class GeoTest {
             .put("US", "US-CA")//
             .put("IM", "IM").build().asMap());
 
-    String asJson = "{\"type\":\"geo\",\"regions\":{\"US\":[\"US-VA\",\"US-CA\"],\"IM\":[\"IM\"]}}";
+    String asJson = "{\"regions\":{\"US\":[\"US-VA\",\"US-CA\"],\"IM\":[\"IM\"]}}";
 
     public void serializeNaturallyAsJson() throws JsonProcessingException {
         assertEquals(new ObjectMapper().writeValueAsString(geo), asJson);
@@ -31,51 +27,5 @@ public class GeoTest {
 
     public void equalToDeserializedMap() throws IOException {
         assertEquals(new ObjectMapper().readValue(asJson, Map.class), geo);
-    }
-
-    static Map<String, Object> geoWhereRegionsAreMapStringCollectionString = ImmutableMap.<String, Object> builder()//
-            .put("type", "geo")//
-            .put("regions", ImmutableMultimap.<String, String> builder()//
-                    .put("US", "US-VA")//
-                    .put("US", "US-CA")//
-                    .put("IM", "IM").build().asMap()).build();
-
-    public void asGeo() {
-        assertEquals(Geo.asGeo(geoWhereRegionsAreMapStringCollectionString), geo);
-    }
-
-    @Test(expectedExceptions = IllegalArgumentException.class, expectedExceptionsMessageRegExp = "expected profile to have a Map<String, Collection<String>> regions field, not class com.google.common.collect.ImmutableListMultimap")
-    public void asGeoNoGuavaAllowed() {
-        Map<String, Object> geoWhereRegionsAreMultimapStringString = ImmutableMap.<String, Object> builder()//
-                .put("type", "geo")//
-                .put("regions", ImmutableMultimap.<String, String> builder()//
-                        .put("US", "US-VA")//
-                        .put("US", "US-CA")//
-                        .put("IM", "IM").build()).build();
-
-        Geo.asGeo(geoWhereRegionsAreMultimapStringString);
-    }
-
-    @Test(expectedExceptions = IllegalArgumentException.class, expectedExceptionsMessageRegExp = "expected regions values to be a subtype of Collection<String>, not String")
-    public void asGeoMultivalueTerritories() {
-        Map<String, Object> geoWhereRegionsAreMultimapStringString = ImmutableMap.<String, Object> builder()//
-                .put("type", "geo")//
-                .put("regions", ImmutableMap.<String, String> builder()//
-                        .put("US", "US-VA")//
-                        .put("IM", "IM").build()).build();
-
-        Geo.asGeo(geoWhereRegionsAreMultimapStringString);
-    }
-
-    static ResourceRecordSet<AData> geoRRS = ResourceRecordSet.<AData> builder()//
-            .name("www.denominator.io.")//
-            .type("A")//
-            .qualifier("US-East")//
-            .ttl(3600)//
-            .add(AData.create("1.1.1.1"))//
-            .addProfile(geoWhereRegionsAreMapStringCollectionString).build();
-
-    public void asGeoRRSet() {
-        assertEquals(Geo.asGeo(geoRRS), geo);
     }
 }